wip: applyReaction directly on tree
