+++ src/reaction/utils/applyOneReactantReaction.js
@@ -1,3 +1,4 @@
+import { checkIfExistsOrAddInfo } from './checkIfExistsOrAddInfo';
 
 /**
  * @description apply one reaction to one reactant
@@ -56,7 +57,6 @@ export function applyOneReactantReaction(reactants, reactions, options) {
                 children: [],
               };
               products.push(product);
-
               todoNextDepth.push(() => {
                 return applyOneReactantReaction([reactionProduct], reactions, {
                   ...options,
@@ -86,40 +86,3 @@ export function applyOneReactantReaction(reactants, reactions, options) {
   return todoNextDepth;
 }
 
-function checkIfExistsOrAddInfo(processedMolecules, molecule, options) {
-  const { moleculeInfoCallback, asReagent, asProduct } = options
-  const idCode = molecule.getIDCode();
-  if (processedMolecules.has(idCode)) {
-    const entry = processedMolecules.get(idCode);
-    let exists = false;
-    if (asReagent) {
-      if (entry.asReagent) {
-        exists = true
-      } else {
-        entry.asReagent = true;
-      }
-    }
-    if (asProduct) {
-      if (entry.asProduct) {
-        exists = true
-      } else {
-        entry.asProduct = true;
-      }
-    }
-    return { exists, info: entry };
-  } else {
-    let info = {
-      idCode,
-      molfile: molecule.toMolfile(),
-      asReagent,
-      asProduct,
-      info: {}
-    }
-    if (moleculeInfoCallback) {
-      info.info = moleculeInfoCallback(molecule)
-    }
-    processedMolecules.set(idCode, info);
-    return { exists: false, info };
-  }
-
-}
